+++ src/pages/api/dashboard/products/index.ts
@@ -14,20 +14,33 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         }
         res.status(200).json(products)
     } else if (req.method === "DELETE") {
-        const { product_uid } = req.headers;
+        const product_uid = req.headers["product_uid"] as string;
+
+        console.log("testttttt " + product_uid)
         if (!product_uid) {
             res.status(400).json({ error: "product uid not specified ! P-001" })
             return;
         }
-        const deleted = await prisma.product.delete({
+        const deletedProductImage = await prisma.productImages.deleteMany({
             where: {
-                product_uid: product_uid as string
+                product_uid: product_uid
             }
         })
-        if (deleted) {
-            res.status(200).json({ message: "the product has been deleted" });
-        } else {
+        if (deletedProductImage) {
+            const deleted = await prisma.product.delete({
+                where: {
+                    product_uid: product_uid
+                }
+            })
+            if (deleted) {
+                res.status(200).json({ message: "the product has been deleted" });
+            } else {
+                res.status(400).json({ error: "Error during deleting. P-002" })
+            }
+        }
+        else {
             res.status(400).json({ error: "Error during deleting. P-002" })
         }
+
     }
 }
